eduro1-2-main 2/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.1.0
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.2
python-decouple==3.8
